requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[argon2]